Refactor of working code
