+++ lib/helpers/endpoints_generator.js
@@ -1,9 +1,10 @@
 var routeGenerator = require('./route_generator.js');
 
-/*
+/**
 * @param {object}
 * @return {array[object]} - return the Hapi routes
 */
+
 module.exports = function(endpoints, html) {
   var routes = [];
   Object.keys(endpoints).forEach(function(endpoint_name) {
