+++ lib/helpers/handler_generator.js
@@ -1,20 +1,15 @@
-var pageToHtml = require('./page_to_html');
+var createContent = require('./html_transform/create_content.js');
 
 module.exports = function (endpoint_name, page, method, html) {
   var path = page.path ? page.path : '/' + endpoint_name;
+
+  var content = createContent(path, page.contents, html);
+
   return function (request, reply) {
     if (method === 'GET') {
-      var content = '<form method="POST" action="' + path + '">';
-      page.contents.forEach(function(tag) {
-        if (tag.options.confirmation) {
-          content += html[tag.field];
-        }
-        content += html[tag.field];
-      });
-      content += '<button type="submit">SUBMIT</button></form>';
       return reply(content);
     } else if (method === 'POST') {
-      return reply('POST html for ' + endpoint_name);
+      return reply('POST request for ' + endpoint_name);
     }
   };
 };
