+++ lib/index.js
@@ -1,17 +1,10 @@
 var endpointsGenerator = require('./helpers/endpoints_generator.js');
-var fieldToHtml = require('./helpers/field_to_html.js');
+var fieldsToHtml = require('./helpers/html_transform/fields_to_html.js');
 
 exports.register = function (server, options, next) {
-
-  var html = {};
-
-  Object.keys(options.fields).forEach(function (field) {
-    html[field] = fieldToHtml(field, options.fields[field]);
-  });
-
-  console.log(html);
-
+  var html = fieldsToHtml(options.fields);
   var routes = endpointsGenerator(options.pages, html);
+
   server.route(routes);
   return next();
 };
