+++ test/helpers/fields_to_html.test.js
@@ -1,5 +1,5 @@
 var tape = require('tape');
-var field_to_html = require('../../lib/helpers/field_to_html.js');
+var fieldToHtml = require('../../lib/helpers/html_transform/field_to_html.js');
 
 tape('returns a html input tag', function (t) {
   var name = 'email';
@@ -9,7 +9,7 @@ tape('returns a html input tag', function (t) {
     value: '',
     required: true
   };
-  var actual = field_to_html(name, input);
+  var actual = fieldToHtml(name, input);
   var expected = '<label>email: </label><input type="string" value="" \>';
   t.equal(actual, expected);
   t.end();
@@ -22,7 +22,7 @@ tape('returns a html select tag', function (t) {
     values: ['M', 'F', 'O', 'NA'],
     value: 'F'
   };
-  var actual = field_to_html(name, select);
+  var actual = fieldToHtml(name, select);
   var expected = '<label>gender: </label>'
     + '<select>'
     + '<option>M</option>'
@@ -39,7 +39,7 @@ tape('returns a input type string with date input', function (t) {
   var input = {
     type: 'date'
   };
-  var actual = field_to_html(name, input);
+  var actual = fieldToHtml(name, input);
   var expected = '<label>dob: </label><input type="date" \>';
   t.equal(actual, expected);
   t.end();
